Fix buildbreak due to missing casts

+++ src/devices/Card/Mifare/MifareCard.cs
@@ -571,7 +571,7 @@ namespace Iot.Device.Card.Mifare
         /// </summary>
         /// <param name="sectorNumber">Input sector number</param>
         /// <returns>The number of blocks for this specific sector</returns>
-        public byte GetNumberBlocks(byte sectorNumber) => sectorNumber < 32 ? 4 : 16;
+        public byte GetNumberBlocks(byte sectorNumber) => sectorNumber < 32 ? (byte)4 : (byte)16;
 
         /// <summary>
         /// Get the number of blocks for a specific sector
@@ -807,7 +807,7 @@ namespace Iot.Device.Card.Mifare
             authOk &= WriteDataBlock(2);
             authOk &= AuthenticateBlockKeyB(keyFormat, 3);
             Data = new byte[] { 0xA0, 0xA1, 0xA2, 0xA3, 0xA4, 0xA5, 0x78, 0x77, 088, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 };
-            Data[9] = Capacity == MifareCardCapacity.Mifare1K ? 0xC1 : 0xC2;
+            Data[9] = Capacity == MifareCardCapacity.Mifare1K ? (byte)0xC1 : (byte)0xC2;
             keyFormat.CopyTo(Data, 10);
             authOk &= WriteDataBlock(3);
             authOk &= AuthenticateBlockKeyB(keyFormat, 4);
